Add playbook to configure lunasa access for barbican
This playbook will check out an ansible role that creates modified
barbican images and creates the relevant secrets needed.

In addition, the playbook modifies the control plane CR to include
the required config to barbican.

You need to call the update-containers role to be able to use the
updated barbican images.